+++ src/gaOptimize.py
@@ -918,3 +918,4 @@ def plotTraces(posTraces: list[tuple[PRED_AR_T, str, str]],
             corYvals = [bottom, midPt, top, top, midPt, bottom]
         ax.fill(corXvals, corYvals, matplotlib.colors.to_hex(corruptorColors[corType]))
     ax.legend()
+# Copyright 2022, 2023, 2024 Charles McAnany. This file is part of BPReveal. BPReveal is free software: You can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 2 of the License, or (at your option) any later version. BPReveal is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with BPReveal. If not, see <https://www.gnu.org/licenses/>.  # noqa  # pylint: disable=line-too-long